refactor: internal cross-platorm signal handling API
