+++ pueue/src/process_helper/windows.rs
@@ -1,8 +1,10 @@
 // We allow color_eyre in here, as this is a module that'll be strictly used internally.
 // As soon as it's obvious that this is code is intended to be exposed to library users, we have to
 // go ahead and replace any `anyhow` usage by proper error handling via our own Error type.
+use color_eyre::eyre::Error;
 use process_wrap::std::ChildWrapper;
 use pueue_lib::Settings;
+use pueue_lib::message::request::Signal as InternalSignal;
 use winapi::{
     shared::{minwindef::FALSE, ntdef::NULL},
     um::{
@@ -18,18 +20,10 @@ use winapi::{
 };
 
 use crate::internal_prelude::*;
+use crate::process_helper::ProcessAction;
 
 /// A handle to a spawned child process.
 type ChildHandle = Box<dyn ChildWrapper>;
-
-/// Shim signal enum for windows.
-pub enum Signal {
-    SIGINT,
-    SIGKILL,
-    SIGTERM,
-    SIGCONT,
-    SIGSTOP,
-}
 
 pub fn get_shell_command(settings: &Settings) -> Vec<String> {
     let Some(ref shell_command) = settings.daemon.shell_command else {
@@ -46,32 +40,53 @@ pub fn get_shell_command(settings: &Settings) -> Vec<String> {
     shell_command.clone()
 }
 
-/// Send a signal to a windows process.
-pub fn send_signal_to_child(child: &mut ChildHandle, signal: Signal) -> Result<()> {
+impl TryFrom<InternalSignal> for ProcessAction {
+    type Error = Error;
+
+    fn try_from(signal: InternalSignal) -> Result<ProcessAction> {
+        match signal {
+            InternalSignal::SigCont => Ok(ProcessAction::Resume),
+            InternalSignal::SigStop => Ok(ProcessAction::Pause),
+            InternalSignal::SigInt | InternalSignal::SigKill | InternalSignal::SigTerm => {
+                bail!("Trying to send unix signal on a windows machine. This isn't supported.")
+            }
+        }
+    }
+}
+
+/// Handle pause/resume actions on processes.
+pub fn handle_process_action(child: &mut ChildHandle, action: ProcessAction) -> Result<()> {
     let pids = get_cur_task_processes(child.id());
     if pids.is_empty() {
         bail!("Process has just gone away");
     }
 
-    match signal {
-        Signal::SIGSTOP => {
+    match action {
+        ProcessAction::Pause => {
             for pid in pids {
                 for thread in get_threads(pid) {
                     suspend_thread(thread);
                 }
             }
         }
-        Signal::SIGCONT => {
+        ProcessAction::Resume => {
             for pid in pids {
                 for thread in get_threads(pid) {
                     resume_thread(thread);
                 }
             }
         }
-        _ => {
-            bail!("Trying to send unix signal on a windows machine. This isn't supported.");
-        }
     }
+
+    Ok(())
+}
+
+/// Send a signal to a windows process.
+/// In practice it doesn't really make sense to handle UNIX signals on Windows, but we're doing a
+/// best-effort here.
+pub fn send_signal_to_child(child: &mut ChildHandle, signal: InternalSignal) -> Result<()> {
+    let action: ProcessAction = signal.try_into()?;
+    handle_process_action(child, action)?;
 
     Ok(())
 }
@@ -255,7 +270,7 @@ pub fn process_exists(pid: u32) -> bool {
 
 #[cfg(test)]
 mod test {
-    use std::{process::Command, thread::sleep, time::Duration};
+    use std::{thread::sleep, time::Duration};
 
     use process_wrap::std::*;
